backend/training_platform/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
mysqlclient==2.1.1
Pillow==10.0.1

# environment
# training_platform_venv